+++ src/main/java/com/telepathicgrunt/the_bumblezone/blocks/EmptyHoneycombBrood.java
@@ -19,7 +19,7 @@ import net.minecraft.util.math.BlockRayTraceResult;
 import net.minecraft.world.World;
 
 
-public class EmptyHoneycombBrood extends DirectionalBlock {
+public class EmptyHoneycombBrood extends ProperFacingBlock {
 
     public EmptyHoneycombBrood() {
         super(Block.Properties.create(Material.CLAY, MaterialColor.ADOBE).hardnessAndResistance(0.5F, 0.5F).sound(SoundType.CORAL));
@@ -36,19 +36,6 @@ public class EmptyHoneycombBrood extends DirectionalBlock {
     public BlockState getStateForPlacement(BlockItemUseContext context) {
         return this.getDefaultState().with(FACING, context.getFace().getOpposite());
     }
-
-
-    @Override
-    public BlockState rotate(BlockState state, Rotation rot) {
-        return state.with(FACING, rot.rotate(state.get(FACING)));
-    }
-
-
-    @Override
-    public BlockState mirror(BlockState state, Mirror mirrorIn) {
-        return state.rotate(mirrorIn.toRotation(state.get(FACING)));
-    }
-
 
     /**
      * MOD COMPAT