+++ src/main/java/com/telepathicgrunt/the_bumblezone/blocks/ProperFacingBlock.java
@@ -1,0 +1,23 @@
+package com.telepathicgrunt.the_bumblezone.blocks;
+
+import net.minecraft.block.BlockState;
+import net.minecraft.block.DirectionalBlock;
+import net.minecraft.util.Mirror;
+import net.minecraft.util.Rotation;
+
+public class ProperFacingBlock extends DirectionalBlock {
+
+    protected ProperFacingBlock(Properties builder) {
+        super(builder);
+    }
+
+    @Override
+    public BlockState rotate(BlockState state, Rotation rot) {
+        return state.with(FACING, rot.rotate(state.get(FACING)));
+    }
+
+    @Override
+    public BlockState mirror(BlockState state, Mirror mirrorIn) {
+        return state.rotate(mirrorIn.toRotation(state.get(FACING)));
+    }
+}