+++ src/main/java/com/telepathicgrunt/the_bumblezone/features/BzFeatures.java
@@ -10,8 +10,6 @@ import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
-import java.util.function.Supplier;
-
 public class BzFeatures {
     // For making ore features replace combs
     // Init at mod startup when adding FEATURES to event bus