+++ src/main/java/com/telepathicgrunt/the_bumblezone/features/decorators/BzPlacements.java
@@ -9,8 +9,6 @@ import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
-import java.util.function.Supplier;
-
 public class BzPlacements {
     public static final DeferredRegister<Placement<?>> DECORATORS = DeferredRegister.create(ForgeRegistries.DECORATORS, Bumblezone.MODID);